add test to all layers

+++ src/main/java/gleb/dresher/AccountShopApi/service/SellerService.java
@@ -1,42 +1,60 @@
 package gleb.dresher.AccountShopApi.service;
 
 import gleb.dresher.AccountShopApi.dto.SellerDTO;
+import gleb.dresher.AccountShopApi.dto.SellerResponse;
 import gleb.dresher.AccountShopApi.entity.Account;
 import gleb.dresher.AccountShopApi.entity.Seller;
 import gleb.dresher.AccountShopApi.repository.AccountRepository;
 import gleb.dresher.AccountShopApi.repository.SellerRepository;
-import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 @Service
-@RequiredArgsConstructor
 public class SellerService {
     private final SellerRepository sellerRepository;
     private final AccountRepository accountRepository;
 
-    public List<Seller> getAllSellers() {
-        return sellerRepository.findAll();
+    public SellerService(SellerRepository sellerRepository, AccountRepository accountRepository) {
+        this.sellerRepository = sellerRepository;
+        this.accountRepository = accountRepository;
     }
 
-    public List<Account> getSellerAccounts(int id) {
-        return sellerRepository.findById(id).get().getAccounts();
+    public ResponseEntity<SellerResponse> getSeller(int id) {
+        Optional<Seller> seller = sellerRepository.findById(id);
+
+        if (seller.isPresent()){
+            SellerResponse sellerResponse = new
+                    SellerResponse(seller.get().getName(),
+                    seller.get().getAccounts());
+            return ResponseEntity.ok(sellerResponse);
+        }
+        return ResponseEntity.notFound().build();
     }
 
-    public Seller addSeller(SellerDTO sellerDTO) {
+    public ResponseEntity<Seller> addSeller(SellerDTO sellerDTO) {
         Seller seller = new Seller(sellerDTO.getName(), sellerDTO.getEmail());
-        return sellerRepository.save(seller);
+        URI location = URI.create("/seller/");
+        return ResponseEntity.created(location).body(sellerRepository.save(seller));
     }
 
-    public void deleteSeller(int id) {
+    public ResponseEntity<Void> deleteSeller(int id) {
+        if (!sellerRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         accountRepository.deleteAllBySellerId(id);
         sellerRepository.deleteById(id);
+        return ResponseEntity.noContent().build();
     }
 
-    public String updateSeller(int id, SellerDTO sellerDTO) {
+    public ResponseEntity<Seller> updateSeller(int id, SellerDTO sellerDTO) {
         if (!sellerRepository.existsById(id)) {
-            return "Seller not found";
+            return ResponseEntity.notFound().build();
         }
         Seller seller = sellerRepository.findById(id).get();
         if (sellerDTO.getName() != null) {
@@ -46,6 +64,6 @@ public class SellerService {
             seller.setEmail(sellerDTO.getEmail());
         }
         sellerRepository.save(seller);
-        return "Seller with id " + id + " updated successfully";
+        return ResponseEntity.ok(seller);
     }
 }
